Review cleanup: extract QBO composite key concern, fix impersonation security, fix N+1
- Extract HasQboInvoiceViaCompositeKey concern from invoice_tracker,
  contributor_adjustment, and adhoc_invoice_tracker (shared qbo_invoice
  getter + validation)
- Gate impersonation on true_admin_user.is_admin? in both current_admin_user
  and impersonating? to prevent session key exploitation by non-admins
- Replace Ruby-level .select in PayCycles discovery with SQL EXISTS subquery
  to eliminate N+1 queries
