+++ app/models/adhoc_invoice_tracker.rb
@@ -1,16 +1,9 @@
 class AdhocInvoiceTracker < ApplicationRecord
+  include HasQboInvoiceViaCompositeKey
+
   belongs_to :project_tracker
   belongs_to :qbo_account
   belongs_to :qbo_invoice, class_name: "QboInvoice", foreign_key: "qbo_invoice_id", primary_key: "qbo_id"
-
-  validate :qbo_invoice_must_live_in_qbo_account
-
-  def qbo_invoice
-    in_memory = association(:qbo_invoice).target
-    return in_memory if in_memory.present?
-    return nil unless qbo_invoice_id && qbo_account_id
-    QboInvoice.find_by(qbo_id: qbo_invoice_id, qbo_account_id: qbo_account_id)
-  end
 
   def display_name
     "#{qbo_invoice.try(:display_name)}"
@@ -44,11 +37,4 @@ class AdhocInvoiceTracker < ApplicationRecord
     end || [])
   end
 
-  private
-
-  def qbo_invoice_must_live_in_qbo_account
-    return if qbo_invoice_id.blank? || qbo_account_id.blank?
-    return if QboInvoice.exists?(qbo_id: qbo_invoice_id, qbo_account_id: qbo_account_id)
-    errors.add(:qbo_invoice_id, "no QboInvoice with qbo_id=#{qbo_invoice_id} exists in qbo_account #{qbo_account_id}")
-  end
 end
