+++ lib/stacks/task_builder/discoveries/pay_cycles.rb
@@ -10,7 +10,7 @@ module Stacks
           PayCycle
             .includes(:pay_stubs, enterprise: { enterprise_admins: :admin_user })
             .where(approved_at: nil)
-            .select { |cycle| cycle.pay_stubs.exists? }
+            .where("EXISTS (SELECT 1 FROM pay_stubs WHERE pay_stubs.pay_cycle_id = pay_cycles.id)")
             .map do |cycle|
               task(
                 subject: cycle,
